bas-db: Refactor IMapperTemplate.select(int) to select(long).

+++ bas/layer-5/bas-db/src/main/java/net/bodz/bas/db/batis/IMapperTemplate.java
@@ -9,7 +9,7 @@ public interface IMapperTemplate<T, C>
 
     List<T> filter(C criteria);
 
-    T select(int id);
+    T select(long id);
 
     int insert(T obj);
